starting to transition to use-package
added an empty use-package statement for each package that was in
package-list and use "use-package-always-ensure" to ensure all
packages are installed.

This can be greatly optimized in the future, but this currently
ensures all packages are installed like before.

auctex and org-plus-contrib required special treatment
auctex does not provide the symbol auctex
org-plus-contrib must be ensured in (use-package org)