finally started to fix my keybindings
The keybindings I came up with over the years work great! They are just a bit
all over the place, for instance `<f6>` for magit is one I started to realize is
too far away for the amount I use it.

And while the new keybinding is longer (3 instead of 1 keys) these keys are
right at my fingertips at all times.

This is going to take some time to re-learn the new keybindings but they are far
more organized. Maybe there will be some changes if I realize that they are not
ideal.

Also removed some completely unused keybindings like recompile and gofmt.